requirements.txt
pandas==1.5.3
numpy==1.23.5
streamlit==1.23.1
scikit-surprise==1.1.3
scikit-learn==1.2.2
st-star-rating==0.0.3